// vanila_redux_todo-list/src/enteties/location/model/actionTypes/locationActionTypes.ts
export enum LocationActionTypes {
  SET_IS_LOADING = "LOCATION/SET_IS_LOADING",
  SET_ERROR = "LOCATION/SET_ERROR",
  SET_LOCATIONS = "LOCATION/SET_LOCATION",
  SET_SINGLE_LOCATION = "LOCATION/SET_SINGLE_LOCATION",
  CLEAR_ALL_LOCATION_META = "LOCATION/CLEAR_ALL_LOCATION_META",
  SET_LOCATION_SEARCH_QUERY = "LOCATION/SET_LOCATION_SEARCH_QUERY",
  SET_LOCATION_PAGE = "LOCATION/SET_LOCATION_PAGE",
  SET_LOCATION_PAGE_FORWARD = "LOCATION/SET_LOCATION_PAGE_FORWARD",
  SET_LOCATION_PAGE_BACK = "LOCATION/SET_LOCATION_PAGE_BACK",
}